% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aenmd.data.gencode.v43-package.R
\docType{package}
\name{aenmd.data.gencode.v43.grch37-package}
\alias{aenmd.data.gencode.v43.grch37}
\alias{aenmd.data.gencode.v43.grch37-package}
\title{aenmd.data.gencode.v43.grch37: Data package for the aenmd package, based on GENCODE version 43}
\description{
Annotation of genetic variants with predicted escape from nonsene mediated decay (NMD) is performed by the aenmd package. This package contains processed transcript information based on GENCODE version 43, the GRCh37 version.
}
\keyword{internal}
