% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_snv_tri}
\alias{EA_snv_tri}
\title{Keys and values containing all STOP-generating SNVs in this package's transcript  set}
\description{
Keys and values containing all STOP-generating SNVs in this package's transcript  set
}
\details{
A \code{triebeard::trie}. Contains the keys for all SNVs that generate 
stop codons. The key format is \code{chr:start}, where start is a zero-padded 9-digit number.
The values are strings, each string contains all ensembl transcript ids for which the SNV 
causes a stop. Transcript ids in the string are separated by the "|" character.
}
\author{
Dennis Kostka
}
\keyword{data}
